+++ addon/mixins/select-box/disableable.js
@@ -7,7 +7,7 @@ export default Mixin.create({
     this._super(...arguments);
 
     if (isPresent(this.disabled)) {
-      set(this, 'isDisabled', this.disabled);
+      set(this, 'isDisabled', Boolean(this.disabled));
     }
 
     if (this.isDisabled) {