+++ tests/integration/components/select-box/clicking-outside-test.js
@@ -17,9 +17,9 @@ module('select-box (clicking outside)', function(hooks) {
 
     await render(hbs`
       <div class="outside"></div>
-      {{#select-box onClickOutside=this.clickedOutside}}
+      <SelectBox @onClickOutside={{this.clickedOutside}}>
         <div class="inside"></div>
-      {{/select-box}}
+      </SelectBox>
     `);
 
     await click('.select-box');